save outline width, value factor 10

+++ lazpaint/utool.pas
@@ -197,7 +197,7 @@ type
     ToolShapeAltitude: integer;
     ToolShapeBorderSize: integer;
     ToolShapeType: string;
-    ToolTextOutlineWidth: integer;
+    ToolTextOutlineWidth: single;
     ToolTextAlign: TAlignment;
     ToolBrushInfoIndex: integer;
     ToolBrushSpacing: integer;
